Moved suggestion trait + cleanup
