+++ src/mako/reactor/Dispatcher.php
@@ -7,11 +7,11 @@
 
 namespace mako\reactor;
 
+use mako\common\traits\SuggestionTrait;
 use mako\reactor\exceptions\InvalidArgumentException;
 use mako\reactor\exceptions\InvalidOptionException;
 use mako\reactor\exceptions\MissingArgumentException;
 use mako\reactor\exceptions\MissingOptionException;
-use mako\reactor\traits\SuggestionTrait;
 use mako\syringe\Container;
 use mako\utility\Str;
